requirements.txt
langchain
langchain-community
langchain-openai
pypdf
pandas
numpy
streamlit
python-dotenv
qdrant-client